ref(spans): Move to separate file
